moved message length checks to Textbox component
